添加一些小功能

+++ manage/src/components/hrl/order.vue
@@ -1,12 +1,11 @@
 <template>
   <div>
-
     <el-table
       ref="multipleTable"
       :data="pageData"
       tooltip-effect="dark"
       style="width: 100%"
-      :default-sort = "{prop: 'price', order: 'descending'}"
+      :default-sort="{prop: 'price', order: 'descending'}"
       @selection-change="handleSelectionChange"
     >
       <el-table-column type="selection" width="55"></el-table-column>
@@ -62,9 +61,9 @@ export default {
   data() {
     return {
       tableData: [],
-      pageData:[],
+      pageData: [],
       multipleSelection: [],
-      currentPage3: 1,
+      currentPage3: 1
     };
   },
   async created() {
@@ -72,7 +71,7 @@ export default {
       data: { data }
     } = await this.$axios.get("http://127.0.0.1:1901/hrl/order");
     this.tableData = data;
-    this.pageData = data.slice(0,4);
+    this.pageData = data.slice(0, 4);
   },
   methods: {
     toggleSelection(rows) {
@@ -88,28 +87,38 @@ export default {
       this.multipleSelection = val;
     },
     async deleter(goods_id) {
-      let {
-        data: { data }
-      } = await this.$axios.delete("http://127.0.0.1:1901/hrl/order", {
-        params: {
-          goods_id: goods_id
-        }
-      });
+      if (this.$store.state.common.delete) {
+        let {
+          data: { data }
+        } = await this.$axios.delete("http://127.0.0.1:1901/hrl/order", {
+          params: {
+            goods_id: goods_id
+          }
+        });
 
-      let newData = await this.$axios.get("http://127.0.0.1:1901/hrl/order")
-      this.tableData = newData.data.data;
-      this.pageData = newData.data.data.slice(0,4);
+        let newData = await this.$axios.get("http://127.0.0.1:1901/hrl/order");
+        this.tableData = newData.data.data;
+        this.pageData = newData.data.data.slice(0, 4);
+      } else {
+        alert("权限不足");
+      }
     },
     async complete(complete, id) {
-      for (let i = 0; i < this.tableData.length; i++) {
-        if (this.tableData[i]._id === id) {
-          this.tableData[i].complete = true;
-          this.$refs.multipleTable.setCurrentRow(id) 
+      if (this.$store.state.common.update) {
+        for (let i = 0; i < this.tableData.length; i++) {
+          if (this.tableData[i]._id === id) {
+            this.tableData[i].complete = true;
+            this.$refs.multipleTable.setCurrentRow(id);
+          }
         }
-      };
-      let {data:{data}} = await this.$axios.patch("http://127.0.0.1:1901/hrl/order",{
-          _id:id
-      })
+        let {
+          data: { data }
+        } = await this.$axios.patch("http://127.0.0.1:1901/hrl/order", {
+          _id: id
+        });
+      } else {
+        alert("权限不足");
+      }
     },
     //分页功能
     handleSizeChange(val) {
@@ -117,13 +126,11 @@ export default {
     },
     handleCurrentChange(val) {
       let newval = 4 * val;
-      this.pageData = this.tableData.slice(newval - 4,newval);
-    },
-    
+      this.pageData = this.tableData.slice(newval - 4, newval);
+    }
   }
 };
 </script>
 
 <style scoped>
-
 </style>
